backend/requirements.txt
fastapi
uvicorn[standard]
neo4j
kafka-python
redis
